Choose a Point-functionality added
By using a custom View class DragPointView, the user can choose and
move a Point on the image. The coordinates of the point are displayed.
The application is now locked in landscape mode.
still todo: make the user add exactly four points and connect them to
quadrangle. Pass the coordinates to the image manipulation section.