added new filters to improve the recognition rate
The algorithm now uses custom edge detection + noise removal +
hough-line-transform do  align the image before applying
the detection mask.

Currently this script DOES NOT APPLY the mask or outputs the lmr
string. The mask needs to be adjusted to the input image size.
Previously the transformed image was of fixed size.